Reinstate the lambda lifting after the CPS transform
This is to avoid too many nested function definitions, which can cause
stack overflows in JS parsers. There are now two lambda lifting passes
in total: one before and one after CPS-transforming the code. The
pre-CPS-transform lifting is done to avoid nesting functions that are
doubly-translated, which would be both costly in code size and incorrect
in terms of closure environment capture.